refactor(runner): extract run-turn.ts from sandbox_agent.ts
Move runTurn whole and unchanged into run-turn.ts. Monolith re-exports it and
imports it back for runSandboxAgent. Pure code move; typecheck clean; 1158 runner
unit tests pass.
